Fix MSVC compiler warnings
reg. 'return': conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/engraving/libmscore/lyrics.cpp
@@ -319,8 +319,8 @@ void Lyrics::layout()
 //                   LOGD("create leading, trailing <%s> -- <%s><%s>", qPrintable(text), qPrintable(leading), qPrintable(trailing));
             const TextBlock& tb = textBlock(0);
 
-            const qreal leadingWidth = tb.xpos(leading.size(), this) - tb.boundingRect().x();
-            const int trailingPos = text.size() - trailing.size();
+            const qreal leadingWidth = tb.xpos(static_cast<int>(leading.size()), this) - tb.boundingRect().x();
+            const int trailingPos = static_cast<int>(text.size() - trailing.size());
             const qreal trailingWidth = tb.boundingRect().right() - tb.xpos(trailingPos, this);
 
             leftAdjust = leadingWidth;
@@ -445,7 +445,7 @@ void Lyrics::paste(EditData& ed, const String& txt)
         }
         underscore = true;
     } else if (sl[0].contains(u"_")) {
-        int p = sl[0].indexOf(u'_');
+        size_t p = sl[0].indexOf(u'_');
         score()->undo(new InsertText(cursorFromEditData(ed), sl[0]), &ed);
         sl[0] = sl[0].mid(p + 1);
         if (sl[0].isEmpty()) {
